rename gnd_pos_control to GroundRoverPositionControl
